Move download btn to <TrendContainer />
