+++ src/components/TrendChart.js
@@ -6,15 +6,12 @@ import { line } from 'd3-shape'
 import { timeParse } from 'd3-time-format'
 import startCase from 'lodash.startcase'
 import throttle from 'lodash.throttle'
-import pluralize from 'pluralize'
 import PropTypes from 'prop-types'
 import React from 'react'
 
-import DownloadDataBtn from './DownloadDataBtn'
 import TrendChartDetails from './TrendChartDetails'
 import XAxis from './XAxis'
 import YAxis from './YAxis'
-import { generateCrimeReadme } from '../util/content'
 import { slugify } from '../util/text'
 
 class TrendChart extends React.Component {
@@ -61,7 +58,6 @@ class TrendChart extends React.Component {
       colors,
       data,
       dispatch,
-      place,
       showMarkers,
       since,
       size,
@@ -79,20 +75,6 @@ class TrendChart extends React.Component {
         ...keysWithSlugs.map(k => ({ [k.slug]: d[k.slug] })),
       ),
     )
-
-    const download = [
-      {
-        data,
-        filename: `${place}-${crime}-${since}–${until}.csv`,
-      },
-      {
-        content: generateCrimeReadme({
-          crime,
-          title: `Reported ${pluralize(crime)} in ${startCase(place)}, ${since}-${until}`,
-        }),
-        filename: 'README.md',
-      },
-    ]
 
     const gaps = []
     const dataByKey = keysWithSlugs.map(k => {
@@ -276,7 +258,6 @@ class TrendChart extends React.Component {
         <div className="my1 fs-10 sm-fs-12 italic monospace center">
           (Crime counts include FBI estimates; rates are rounded)
         </div>
-        <DownloadDataBtn data={download} />
       </div>
     )
   }
